make the delete test and check linters
