+++ src/modules/__mock__/listMock.js
@@ -7,16 +7,10 @@ export const addTaskMock = (task) => {
   localStorage.setItem('tasks', toDoListArray);
 };
 
-export const deleteTask = (e, toDoListArray) => {
-  // const clickedCross = e.target.closest('.trash-can');
-  // const clickedTask = clickedCross.previousElementSibling;
-  // const taskIndex = toDoListArray.findIndex(
-  //   (task) => task.task === clickedTask.value
-  // );
-  // toDoListArray.splice(taskIndex, 1);
-  // toDoListArray.forEach((task, index) => {
-  //   task.id = index + 1;
-  // });
-  // updateLocalStorage(toDoListArray);
-  // renderToDoList(toDoListArray);
+export const deleteTaskMock = (taskIndex) => {
+  toDoListArray.splice(taskIndex, 1);
+  toDoListArray.forEach((task, index) => {
+    task.id = index + 1;
+  });
+  localStorage.setItem('tasks', toDoListArray);
 };
